A file's time is stamped as sixty-four bits: one standard library counts it in a hundred and twenty-eight, which to_string cannot take
The macOS lane could not compile the stamp the themes folder keeps of each theme it has converted: libc++'s file_time_type counts in __int128, for which std::to_string is ambiguous. libstdc++ and MinGW count in sixty-four bits, so no other lane saw it.

+++ src/main.cpp
@@ -1540,7 +1540,9 @@ bool convert_dropped_themes(std::string const& directory)
         std::filesystem::path const stamp = converted / (entry.path().filename().string() + ".stamp");
         std::filesystem::file_time_type const changed = std::filesystem::last_write_time(entry.path(), error);
         std::optional<std::string> const stamped = read_text_file(stamp);
-        std::string const now = std::to_string(changed.time_since_epoch().count());
+        // As sixty-four bits: one standard library counts a file's time in
+        // a hundred and twenty-eight, which to_string has no one answer for.
+        std::string const now = std::to_string(static_cast<long long>(changed.time_since_epoch().count()));
         if (stamped && *stamped == now)
             continue;
         std::vector<std::string> problems;
